refactor: namespae van states in Article.php
